requirements.txt
interactions
PyYAML
requests
websockets